requirements.txt
pandas
PyYAML >= 6.0, < 7.0
tabulate >= 0.9.0, < 1.0
